pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "matrix-webhook-bridge"
version = "0.4.0"
description = "A webhook-to-Matrix notification bridge with per-sender bot users"
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.115,<1.0",
    "jsonschema>=4.0,<5.0",
    "prometheus-client>=0.20,<1.0",
    "pyyaml>=6.0,<7.0",
    "uvicorn[standard]>=0.30,<1.0",
]

[project.scripts]
matrix-webhook-bridge = "matrix_webhook_bridge.cli:main"

[project.optional-dependencies]
test = ["pytest>=8", "httpx>=0.27"]

[tool.bandit]
exclude_dirs = ["tests"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "F",   # pyflakes
  "I",   # isort
  "UP",  # pyupgrade
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
